add popup thing for the link on mobile + delete the stupid div and make it part of header
