+++ src/components/Header.jsx
@@ -1,4 +1,4 @@
-import { useContext } from "react";
+import { useState, useEffect, useContext } from "react";
 import { useRouter } from "next/router";
 import Link from "next/link";
 import { DisplayNameContext } from "@/context/DisplayNameContext";
@@ -7,6 +7,9 @@ import HorizontalDivider from "@/components/HorizontalDivider";
 import styles from "@/styles/components/header.module.css";
 
 export default function Header() {
+  const [windowWidth, setWindowWidth] = useState(0);
+  const [isOpen, setIsOpen] = useState(false);
+
   const { displayName, setDisplayName } = useContext(DisplayNameContext);
 
   const localDate = new Date().toLocaleDateString("en-UK");
@@ -19,8 +22,40 @@ export default function Header() {
     setDisplayName("");
   };
 
+  const handleOverlay = () => {
+    setIsOpen(!isOpen);
+  };
+
+  useEffect(() => {
+    setWindowWidth(window.innerWidth);
+    const handleResize = () => {
+      setWindowWidth(window.innerWidth);
+    };
+
+    window.addEventListener("resize", handleResize);
+    return () => window.removeEventListener("resize", handleResize);
+  }, []);
+
   return (
     <header className={styles.header}>
+      {windowWidth <= 1200 && (
+        <>
+          <img
+            className={styles["overlay-icon"]}
+            onClick={handleOverlay}
+            src={isOpen ? "/x-icon.png" : "/3-lines-icon.png"}
+          />
+
+          <nav
+            className={`${styles.overlay} ${isOpen ? styles.overlayOpen : ""}`}
+          >
+            <Link className={styles["overlay-link"]} href="/">
+              pokédex
+            </Link>
+          </nav>
+        </>
+      )}
+
       <div className={styles["logo-and-user-information"]}>
         <div className={styles.logo}>
           <img src="/pokeball.png" />
@@ -41,19 +76,22 @@ export default function Header() {
               <HorizontalDivider />
             </div>
 
-            <div className={styles.links}>
-              <Link className={styles["header-link"]} href="/">
-                pokédex
-              </Link>
-              <Link className={styles["header-link"]} href="/guess-pokemon">
-                guess that pokemon
-              </Link>
-            </div>
+            {windowWidth > 1200 && (
+              <div className={styles.links}>
+                <Link className={styles["header-link"]} href="/">
+                  pokédex
+                </Link>
+              </div>
+            )}
           </>
         )}
       </div>
 
-      <p className={styles["display-date"]}>{localDate}</p>
+      {windowWidth > 1200 && (
+        <>
+          <p className={styles["display-date"]}>{localDate}</p>
+        </>
+      )}
     </header>
   );
 }
