+++ src/styles/components/header.module.css
@@ -9,6 +9,50 @@
 
   background-color: rgba(220, 220, 240, 0.6);
 
+  & .overlay {
+    position: fixed;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    gap: 36px;
+
+    height: 0;
+    width: 100%;
+
+    background-color: rgb(100, 100, 200);
+
+    overflow-x: hidden;
+
+    z-index: 2;
+
+    transition: height 0.3s ease;
+
+    & .overlay-link {
+      color: rgb(255, 255, 255);
+    }
+  }
+
+  & .overlayOpen {
+    height: 30%;
+
+    transition: height 0.3s ease;
+  }
+
+  & .overlay-icon {
+    position: absolute;
+    top: 8px;
+    right: 12px;
+
+    height: 64px;
+    aspect-ratio: 1 / 1;
+
+    cursor: pointer;
+
+    z-index: 3;
+  }
+
   & .logo-and-user-information {
     display: flex;
     align-items: center;
@@ -57,5 +101,115 @@
 
   & p {
     font-size: 1.8rem;
+  }
+}
+
+@media only screen and (max-width: 1200px) {
+  .header {
+    height: 30%;
+
+    justify-content: center;
+
+    flex-direction: column;
+
+    & .logo-and-user-information {
+      flex-direction: column;
+
+      & .horizontal-divider-container {
+        height: 0;
+      }
+
+      & .username-and-logout {
+        display: flex;
+        align-items: baseline;
+        gap: 16px;
+
+        & button {
+          padding: 8px;
+
+          border-radius: 8px;
+
+          font-size: 2.5rem;
+        }
+      }
+
+      & .links {
+        display: flex;
+        gap: 24px;
+
+        padding-inline: 12px;
+        & .header-link {
+          font-size: 2rem;
+          text-decoration: none;
+        }
+      }
+    }
+
+    & .overlay {
+      position: fixed;
+
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      flex-direction: column;
+      gap: 36px;
+
+      height: 0;
+      width: 100%;
+
+      background-color: rgb(100, 100, 200);
+
+      z-index: 2;
+
+      transition: height 0.3s ease;
+
+      & .overlay-link {
+        color: rgb(255, 255, 255);
+
+        font-size: 5rem;
+      }
+    }
+
+    & .overlayOpen {
+      height: 30%;
+
+      transition: height 0.3s ease;
+    }
+
+    & .logo {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      gap: 12px;
+
+      padding-block: 24px;
+
+      & h1 {
+        font-size: 4.5rem;
+      }
+
+      & img {
+        height: 120px;
+        aspect-ratio: 1 / 1;
+      }
+    }
+
+    & .username-and-logout {
+      display: flex;
+      justify-content: center;
+      align-items: baseline;
+      gap: 16px;
+
+      & p {
+        font-size: 3.5rem;
+      }
+
+      & button {
+        padding: 4px;
+
+        border-radius: 12px;
+        font-size: 2.5rem;
+      }
+    }
   }
 }
