--- src/styles/pages/guess-pokemon.module.css
@@ -1,100 +1,0 @@
-.guess-pokemon-wrapper {
-  height: 100%;
-  width: 100%;
-
-  font-size: 3rem;
-
-  & .mobile-title {
-    height: 30%;
-
-    padding-bottom: 16px;
-
-    overflow-x: hidden;
-
-    & .overlay {
-      position: fixed;
-
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      flex-direction: column;
-      gap: 36px;
-
-      height: 0;
-      width: 100%;
-
-      background-color: rgb(100, 100, 200);
-
-      overflow-x: hidden;
-
-      z-index: 2;
-
-      transition: height 0.3s ease;
-
-      & .overlay-link {
-        color: rgb(255, 255, 255);
-      }
-    }
-
-    & .overlayOpen {
-      height: 30%;
-
-      transition: height 0.3s ease;
-    }
-
-    & .overlay-button {
-      position: absolute;
-      top: 4px;
-      right: 24px;
-
-      border: none;
-
-      color: rgb(255, 255, 255);
-      background-color: transparent;
-
-      z-index: 3;
-      font-size: 4rem;
-    }
-
-    & .logo {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 12px;
-
-      padding-block: 24px;
-
-      & h1 {
-        font-size: 4.5rem;
-      }
-
-      & img {
-        height: 120px;
-        aspect-ratio: 1 / 1;
-      }
-    }
-
-    & .username-and-logout {
-      display: flex;
-      justify-content: center;
-      align-items: baseline;
-      gap: 16px;
-
-      & p {
-        font-size: 3.5rem;
-      }
-
-      & button {
-        padding: 4px;
-
-        border-radius: 12px;
-        font-size: 2.5rem;
-      }
-    }
-  }
-
-  & .p {
-    font-size: 5rem;
-    color: rgb(69, 69, 69);
-  }
-}
